requirements.txt
Django==2.2
djangorestframework==3.12.2
drf-rw-serializers==1.0.4
pytz==2020.4
sqlparse==0.4.1
